requirements.txt
fastapi 
jinja2 
uvicorn 
python-multipart
python-dotenv
